% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electre1.r
\name{electre1}
\alias{electre1}
\title{Electre I}
\usage{
electre1(matrix, weights)
}
\arguments{
\item{matrix}{matrix (n \times m) of n alternatives and m crierias}

\item{weights}{vector of weights size n}

\item{c}{}

\item{d}{}
}
\description{
electre mcdm method
}
