% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSA.R
\name{WSA}
\alias{WSA}
\title{WSA}
\usage{
WSA(matrix, weights, min)
}
\arguments{
\item{matrix}{matrix (n \times m) of n alternatives and m crierias}

\item{weights}{vector of weights size n}

\item{min}{logical vector for which criterias should be minimalized}
}
\description{
wsa mcdm method
}
