% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electre1.r
\name{electre1}
\alias{electre1}
\title{Electre I}
\usage{
electre1(matrix, weights, c, d)
}
\arguments{
\item{matrix}{data.frame (n \times m) of n alternatives and m crierias}

\item{weights}{data.frame of weights size 1 \times n}

\item{c}{}

\item{d}{}
}
\description{
electre mcdm method
}
