% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topsin.r
\name{topsin}
\alias{topsin}
\title{Topsin}
\usage{
topsin(matrix, weights)
}
\arguments{
\item{matrix}{matrix (n \times m) of n alternatives and m crierias}

\item{weights}{vector of weights size n}
}
\description{
electre mcdm method
}
