% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topsis.r
\name{topsis}
\alias{topsis}
\title{Topsis}
\usage{
topsis(matrix, weights)
}
\arguments{
\item{matrix}{data.frame (n \times m) of n alternatives and m crierias}

\item{weights}{data.frame of weights size 1 \times n}
}
\description{
electre mcdm method
}
