% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSA.R
\name{WSA}
\alias{WSA}
\title{WSA}
\usage{
WSA(matrix, weights, which_is_min)
}
\arguments{
\item{matrix}{data.frame (n \times m) of n alternatives and m crierias}

\item{weights}{data.frame of weights size 1 \times n}

\item{which_is_min}{logical vector for which criterias should be minimalized}
}
\description{
wsa mcdm method
}
